Fixes for review, fix parsing lol_us
